complete crud functionality with authorization and authentication on user posts

+++ backend/controllers/postController.js
@@ -1,11 +1,12 @@
 const asyncHandler = require("express-async-handler");
 const Post = require("../models/postModel");
+const User = require("../models/userModel");
 
 // @desc    Get all posts
 // @route   GET /api/posts
 // @access  Private
 const getPosts = asyncHandler(async (req, res) => {
-  const posts = await Post.find({});
+  const posts = await Post.find({ user: req.user.id });
 
   res.status(200).json(posts);
 });
@@ -22,6 +23,7 @@ const createPost = asyncHandler(async (req, res) => {
   const post = await Post.create({
     title: req.body.title,
     text: req.body.text,
+    user: req.user.id,
   });
 
   res.status(200).json(post);
@@ -36,6 +38,20 @@ const updatePost = asyncHandler(async (req, res) => {
   if (!post) {
     res.status(404);
     throw new Error("Post not found");
+  }
+
+  const user = await User.findById(req.user.id);
+
+  // Check if user exists
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+
+  // Check if the user is the owner of the post
+  if (post.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not authorized");
   }
 
   const updatedPost = await Post.findByIdAndUpdate(
@@ -59,6 +75,20 @@ const deletePost = asyncHandler(async (req, res) => {
   if (!post) {
     res.status(404);
     throw new Error("Post not found");
+  }
+
+  const user = await User.findById(req.user.id);
+
+  // Check if user exists
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+
+  // Check if the user is the owner of the post
+  if (post.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not authorized");
   }
 
   await post.deleteOne();
